Don't construct and IDMEF message if object list is empty. We'll do it as soon as we'll have an object to set within it.
Don't check whether the current rule object list is empty to decide whether or
not to send the message. Rather use state->idmef, which will be allocated in case there is something to send. This fix a bug when a top level rule were not setting any object, but was using goto.



git-svn-id: file:///home/yoann/dev/prelude/git/nok/SVN/trunk/prelude-lml@4869 09c5ec92-17d4-0310-903a-819935f44dba