Context / State handling improvement
Theses change fixes a long standing problem with multiple line log
where we would only store information about the first matched log
entry in the IDMEF message. As of now, all of the matched log entry
are available in generated events, as well as all of the rules that
matched and resulted in this alert being generated.