File notification improvement
Always gather our own statistics using fstat() instead of relying on
libev stats that often are outdated.

When we notice a file deletion, trigger reading of any unread events
on the deleted file for which we still have an open file descriptor.

Additionaly, improve handling of fast delete + re-create event.