Initial TCP/TCP-TLS support, various improvement.
* TCP server support has been added and can be activated using the
tcp-server option, the default port is 514.

* TCP-TLS server support has been added and can be activated using the
"tcp-tls-server" option, the defaut port is 6514. The system support
anonymous authentication using the (anon-(EC)DH) key exchange protocol as
well as x509 authentication, which might be used with or without client
certificate verification ("tls-verify option"). Certificate verification
may include fingerprint (MD5 or SHA1) and/or name verification with the
use of the "tls-trusted-fingerprint" and "tls-trusted-name" option.

As part of TCP-TLS development, support for Diffie-Hellman parameters has
been implemented. The parameters are re-generated once a day, and default
to 1024 bits. This might be changed using the "dh-parameters-regenerate"
and "dh-prime-length" options.

Multiple TCP(/TLS) server might be configured, listening on different port
with different options.

The TCP(/TLS) implementation support the two RFC 6587 described method for
reading TCP syslog message :
 - Non transparent framing (raw message terminated with a TRAILER character).
 - Octet counting (message length explicitly indicated by the sender).

* Uniformisation of file length handling. Previously UDP syslog message were
limited to 1024 characters, and message read from file limited to 65535
characters. The input limit is now configurable through a command line and
a configuration option (log-max-length). The default limit has been adjusted
from 1024 to 8192 characters, which is the biggest value from the various
syslog RFC:
 - RFC 3164 (BSD Syslog protocol): packet MUST be 1024 bytes or less.
 - RFC 5424 (The Syslog protocol), section 6.1: messages of up to and
   including 2048 octets in length.  Transport receivers MAY receive
   messages larger than 2048 octets in length.
 - RFC 5425 (TLS Transport Mapping for Syslog), section 4.3.1: SHOULD
   be able to process messages with lengths up to and including 8192 octets.
 - RFC 5426 (Transmission of Syslog Messages over UDP), section 3.2:
   message sizes of up to and including 2048 octets. The ability to receive
   larger messages is encouraged.

* A common, global log buffer is now used by the UDP, TCP and TCP-TLS server
implementation, instead of creating a new buffer on the stack each time
the receiving function are called.

Change-Id: I5c07d2800eed833c937c6bf9d31e90b36ad16521